refactor(`StateContainer`): change the `StateStorage` to `StateContainer` and use `Container`.
